Refactor and consolidate cross-shard FP8 scale handling tests
- Removed the test file `test_model_free_cross_shard.py` and integrated its tests into `test_model_free_utils.py`.
- Added helper functions for writing fake FP8 shards and index JSON files.
- Consolidated tests for `_build_cross_shard_pairs_from_weight_map` and `_hydrate_missing_fp8_scales_from_index` into a single test class.
- Enhanced test coverage for cross-shard scenarios, including hydration in both local and streaming modes.
- Updated imports and ensured compatibility with existing utilities.
